quick patch home screen code cleanup
