+++ src/main/kotlin/app/morphe/gui/ui/screens/quick/components/IdleContent.kt
@@ -1,0 +1,107 @@
+/*
+ * Copyright 2026 Morphe.
+ * https://github.com/MorpheApp/morphe-desktop
+ */
+
+package app.morphe.gui.ui.screens.quick.components
+
+import app.morphe.gui.ui.icons.MorpheIcons
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawBehind
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import app.morphe.gui.ui.theme.*
+
+// ============================================================================
+// IDLE CONTENT (drop prompt)
+// ============================================================================
+
+@Composable
+internal fun IdleContent(
+    isAnalyzing: Boolean,
+    isDragHovering: Boolean,
+    onBrowse: () -> Unit
+) {
+    val corners = LocalMorpheCorners.current
+    val accents = LocalMorpheAccents.current
+    val bracketColor = if (isDragHovering) accents.primary.copy(alpha = 0.7f)
+        else MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.15f)
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .clickable(enabled = !isAnalyzing) { onBrowse() }
+            .drawBehind {
+                val strokeWidth = 2f
+                val len = 32.dp.toPx()
+                val inset = 0f
+
+                // Top-left
+                drawLine(bracketColor, Offset(inset, inset), Offset(inset + len, inset), strokeWidth)
+                drawLine(bracketColor, Offset(inset, inset), Offset(inset, inset + len), strokeWidth)
+                // Top-right
+                drawLine(bracketColor, Offset(size.width - inset, inset), Offset(size.width - inset - len, inset), strokeWidth)
+                drawLine(bracketColor, Offset(size.width - inset, inset), Offset(size.width - inset, inset + len), strokeWidth)
+                // Bottom-left
+                drawLine(bracketColor, Offset(inset, size.height - inset), Offset(inset + len, size.height - inset), strokeWidth)
+                drawLine(bracketColor, Offset(inset, size.height - inset), Offset(inset, size.height - inset - len), strokeWidth)
+                // Bottom-right
+                drawLine(bracketColor, Offset(size.width - inset, size.height - inset), Offset(size.width - inset - len, size.height - inset), strokeWidth)
+                drawLine(bracketColor, Offset(size.width - inset, size.height - inset), Offset(size.width - inset, size.height - inset - len), strokeWidth)
+            },
+        contentAlignment = Alignment.Center
+    ) {
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+            if (isAnalyzing) {
+                CircularProgressIndicator(
+                    modifier = Modifier.size(40.dp),
+                    color = accents.primary,
+                    strokeWidth = 3.dp
+                )
+                Spacer(modifier = Modifier.height(16.dp))
+                Text(
+                    text = "Analyzing APK…",
+                    fontSize = 15.sp,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            } else {
+                Icon(
+                    imageVector = MorpheIcons.CloudUpload,
+                    contentDescription = null,
+                    modifier = Modifier.size(44.dp),
+                    tint = if (isDragHovering) accents.primary
+                           else MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.5f)
+                )
+                Spacer(modifier = Modifier.height(16.dp))
+                Text(
+                    text = "Drop APK here",
+                    fontSize = 17.sp,
+                    fontWeight = FontWeight.SemiBold,
+                    color = if (isDragHovering) accents.primary
+                           else MaterialTheme.colorScheme.onSurface
+                )
+                Spacer(modifier = Modifier.height(4.dp))
+                Text(
+                    text = "or click to browse",
+                    fontSize = 13.sp,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.5f)
+                )
+                Spacer(modifier = Modifier.height(6.dp))
+                Text(
+                    text = ".apk  ·  .apkm  ·  .xapk  ·  .apks",
+                    fontSize = 11.sp,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.3f)
+                )
+            }
+        }
+    }
+}
+
